RmsDetection and BucketScanner now have same constructor. Next step is to remove "external" attack and release and move it to applications.

+++ emmef-parent/noise/src/main/java/org/emmef/audio/noisereduction/DefaultTimings.java
@@ -68,7 +68,7 @@ public class DefaultTimings implements Timings {
 		return lowestFrequency;
 	}
 	
-	public final Timings withSampleRate(double sampleRate) {
+	public final Timings withSampleRate(long sampleRate) {
 		return new RatedTimings(this, sampleRate);
 	}
